[WB-4336] Annotate data_types.py (Part 1: 2286 lines converted, 1275 to go ) (#1764)
