+++ mypy.ini
@@ -83,6 +83,17 @@ check_untyped_defs = True
 disallow_untyped_decorators = True
 strict_equality = True
 
+[mypy-wandb.sdk.data_types]
+disallow_incomplete_defs = True
+disallow_untyped_defs = True
+warn_unused_ignores = True
+warn_return_any = True
+warn_unreachable = True
+check_untyped_defs = True
+# disallow_untyped_calls = True
+disallow_untyped_decorators = True
+strict_equality = True
+
 [mypy-wandb.sdk.lib.telemetry]
 disallow_untyped_defs = True
 disallow_untyped_calls = True
