+++ wandb/apis/public.py
@@ -2793,7 +2793,7 @@ class Artifact(object):
             with open(item_path, "r") as file:
                 json_obj = json.load(file)
             result = wb_class.from_json(json_obj, self)
-            result.artifact_source = {"artifact": self, "name": name}
+            result.set_artifact_source(self, name)
             return result
 
     def download(self, root=None, recursive=False):
