+++ wandb/sdk/internal/run.py
@@ -5,7 +5,7 @@
 Semi-stubbed run for internal process use.
 
 """
-from wandb import data_types
+from wandb._globals import _datatypes_set_callback
 
 from .. import wandb_run
 
@@ -17,7 +17,7 @@ class InternalRun(wandb_run.Run):
 
         # TODO: This overwrites what's done in the constructor of wandb_run.Run.
         # We really want a common interface for wandb_run.Run and InternalRun.
-        data_types._datatypes_set_callback(datatypes_cb)
+        _datatypes_set_callback(datatypes_cb)
 
     def _set_backend(self, backend):
         # This type of run object can't have a backend
